refactor: move the decoder to a dedicated page
This refactor introduced the `DecoderContextProvider`
