+++ src/pages/Decoder/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Decoder.tsx"
+}
